requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
numpy==2.4.2
pandas==3.0.1
scikit-learn==1.8.0
openai==1.3.7
slowapi==0.1.9
psycopg2-binary==2.9.9
httpx==0.25.2
itsdangerous==2.1.2
resend==2.10.0